Add IR de-serializer util
