+++ src/frontends/ir/src/CMakeLists.txt
@@ -5,6 +5,7 @@
 ov_add_frontend(NAME ir
                 FILEDESCRIPTION "FrontEnd to load OpenVINO IR file format"
                 LINK_LIBRARIES openvino::pugixml
-                               openvino::core::dev)
+                               openvino::core::dev
+                               openvino::xml_util)
 
 ov_build_target_faster(openvino_ir_frontend PCH)
